+++ Tryouts/Messaging-JS/src/exceptions/ErrorTypes.ts
@@ -1,0 +1,9 @@
+// TODO: Canonize standard error types
+export const ErrorTypes = {
+    default: "MorganStanley.ComposeUI.Exceptions.MessageRouterException",
+    duplicateEndpoint: "MorganStanley.ComposeUI.Exceptions.DuplicateEndpointException",
+    duplicateRequestId: "MorganStanley.ComposeUI.Exceptions.DuplicateRequestIdException",
+    invalidEndpoint: "MorganStanley.ComposeUI.Exceptions.InvalidEndpointException",
+    invalidTopic: "MorganStanley.ComposeUI.Exceptions.InvalidTopicException",
+    unknownEndpoint: "MorganStanley.ComposeUI.Exceptions.UnknownEndpointException",
+}
