+++ Tryouts/Messaging-JS/src/exceptions/InvalidEndpointError.ts
@@ -1,5 +1,6 @@
 import * as protocol from "../protocol";
-import { ErrorTypes, MessageRouterError } from ".";
+import { ErrorTypes } from "./ErrorTypes";
+import { MessageRouterError } from "./MessageRouterError";
 
 export class InvalidEndpointError extends MessageRouterError {
     constructor(err: { endpoint: string; } | protocol.Error) {
