+++ Tryouts/Messaging-JS/src/exceptions/MessageRouterError.ts
@@ -1,5 +1,5 @@
 import * as protocol from "../protocol";
-import { ErrorTypes } from ".";
+import { ErrorTypes } from "./ErrorTypes";
 
 export class MessageRouterError extends Error {
 
@@ -17,5 +17,27 @@ export class MessageRouterError extends Error {
 
     static fromProtocolError(error: protocol.Error): MessageRouterError {
         throw new Error("Not implemented");
+    }
+}
+
+export function createProtocolError(err: any): protocol.Error {
+    
+    if (err instanceof MessageRouterError) {
+        return {
+            type: err.type ?? ErrorTypes.default,
+            message: err.message
+        }
+    }
+
+    if (err instanceof Error) {
+        return {
+            type: err.name,
+            message: err.message
+        }
+    }
+    
+    return {
+        type: "Error",
+        message: typeof err === "string" ? err : JSON.stringify(err) // TODO: not sure if this is the right way of stringifying an error
     }
 }
