+++ Tryouts/Messaging-JS/src/exceptions/UnknownEndpointError.ts
@@ -1,5 +1,6 @@
 import * as protocol from "../protocol";
-import { ErrorTypes, MessageRouterError } from ".";
+import { ErrorTypes } from "./ErrorTypes";
+import { MessageRouterError } from "./MessageRouterError";
 
 export class UnknownEndpointError extends MessageRouterError {
     constructor(err: { endpoint: string; } | protocol.Error) {
