+++ Tryouts/Messaging-JS/src/protocol/messages/AbstractRequest.ts
@@ -1,4 +1,4 @@
-import { Message } from ".";
+import { Message } from "./Message";
 
 export interface AbstractRequest<TResponse> extends Message {
     readonly requestId: string;
