+++ Tryouts/Messaging-JS/src/protocol/messages/ConnectRequest.ts
@@ -1,4 +1,4 @@
-import { Message } from ".";
+import { Message } from "./Message";
 
 export interface ConnectRequest extends Message {
     type: "Connect";
