+++ Tryouts/Messaging-JS/src/protocol/messages/Message.ts
@@ -1,4 +1,4 @@
-import { MessageType } from ".";
+import { MessageType } from "./MessageType";
 
 export interface Message {
     readonly type: MessageType;
