+++ Tryouts/Messaging-JS/src/protocol/messages/RegisterServiceRequest.ts
@@ -1,5 +1,7 @@
-import { AbstractRequest, Message, RegisterServiceResponse } from ".";
 import { EndpointDescriptor } from "../..";
+import { AbstractRequest } from "./AbstractRequest";
+import { Message } from "./Message";
+import { RegisterServiceResponse } from "./RegisterServiceResponse";
 
 export interface RegisterServiceRequest extends AbstractRequest<RegisterServiceResponse> {
     type: "RegisterService";
