+++ Tryouts/Messaging-JS/src/protocol/messages/RegisterServiceResponse.ts
@@ -1,4 +1,5 @@
-import { Message, AbstractResponse } from ".";
+import { AbstractResponse } from "./AbstractResponse";
+import { Message } from "./Message";
 
 export interface RegisterServiceResponse extends AbstractResponse {
     type: "RegisterServiceResponse";
