+++ Tryouts/Messaging-JS/src/protocol/messages/SubscribeMessage.ts
@@ -1,4 +1,4 @@
-import { Message } from ".";
+import { Message } from "./Message";
 
 export interface SubscribeMessage extends Message {
     type: "Subscribe";
